autotools: change logic for avoiding false cross-compilation in autotools components
Previously, we would force the --host= string to be the same as
the --build= string if the given ISA was supported by the host
system.  This allows run tests in the autoconf script to work.
However, on some systems (e.g. HP-UX hppa64), autoconf needs to
know the true host system type in order to choose the correct
version of programs such as ar.

Instead, force the --build= string to equal the host string.