mk: work around race conditions in mk_mkdir
On some systems, mkdir -p will fail if it races with
another instance to create the same directory.  Work
around this by backing off into an explicit loop to
create each component in the path.  If creating a
component fails but it actually exists as a directory
afterward, assume it was a race and keep going.