mk: add mk_fnmatch_compile function
This takes a list of glob patterns and uses eval to
generate a predicate function that returns true if
the string passed to it matches any of them.  This
can be used to optimize a list of mk_fnmatch tests
into a single function where the shell does most
of the work.