compiler: change strategy for building fat binaries
Rather than emitting multiple targets for each ISA,
emit only one and make the compiler/link scripts
iterate instead.  This only reduces parallelism
if the number of source files is very small,
in which case it's not very important.

This significantly reduces the size of Makefiles
generated on Darwin, which speeds up running
configure and reduces the time it takes make
to parse the file.